requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
email-validator==2.0.0.post1
qrcode==7.4
Pillow==10.0.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
